Finishing example to manually build a portfolio with messy data, and renaming the file from "example-from-files.py" to "Example-Manually-build-Portfolio-with-messy-data.py".
